Rename discord bot package
